:heavy_minus_sign: removes getMyIP() from server
